docs(geoms): resolve illustrations on the server for thin client pages
Move gallery preview lookup into +page.server.ts so detail routes stay free
of EXAMPLES and geom-thumbnails catalogs. Tighten the client-thin guard and
satisfy oxlint consistent-return / no-useless-undefined on the helper.
