+++ apps/docs/src/routes/reference/geoms/[name]/+page.server.ts
@@ -3,6 +3,7 @@ import { error } from "@sveltejs/kit";
 import { GEOM_REFERENCE, type GeomName, KNOWN_GEOMS } from "@ggsvelte/spec";
 
 import { EXAMPLES } from "$lib/examples-manifest";
+import { illustrationForGeom } from "$lib/geom-thumbnails";
 
 import type { EntryGenerator, PageServerLoad } from "./$types";
 
@@ -21,6 +22,22 @@ function relatedExamples(geom: GeomName) {
   ).slice(0, 8);
 }
 
+/** Resolve gallery illustration on the server so the client stays catalog-free. */
+function illustrationData(geom: GeomName, component: string) {
+  const resolved = illustrationForGeom(geom);
+  if (resolved === undefined) {
+    return null;
+  }
+  const example = EXAMPLES.find((ex) => ex.id === resolved.exampleId);
+  return {
+    path: resolved.path,
+    exampleId: resolved.exampleId,
+    title: example?.title ?? component,
+    width: example?.vrWidth ?? 640,
+    height: example?.vrHeight ?? 400,
+  };
+}
+
 export const load: PageServerLoad = ({ params }) => {
   const name = params.name;
   if (!GEOM_SET.has(name)) {
@@ -29,6 +46,7 @@ export const load: PageServerLoad = ({ params }) => {
   const entry = GEOM_REFERENCE[name as GeomName];
   return {
     entry,
+    illustration: illustrationData(entry.name, entry.component),
     examples: relatedExamples(entry.name).map((ex) => ({
       id: ex.id,
       title: ex.title,
